Renamed program.py to device.py
