+++ src/SerialPlotter/device.py
